pyproject.toml
[tool.poetry]
name = "carcalcbot"
version = "0.1.0"
description = ""
authors = ["akm77 <aleksey.kotryakhov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^3.0.0b7"
aiogram-dialog = "^2.0.0b3"
redis = "^4.5.1"
pydantic = "^1.10.6"
python-dotenv = "^1.0.0"
SQLAlchemy = "^2.0.6"
tenacity = "^8.2.2"
aiosqlite = "^0.18.0"
emoji-country-flag = "^1.3.2"
fake-useragent = "^1.1.3"
lxml = "^4.9.2"
apscheduler = "^3.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

